Visualization of Linear Regression overfitting
This code shows the effect of overfitting deriving from adding new sets of features to the database.
Overfitting is described here by the ration between train error and test error. Errors are calculated as average RMSE from 10-fold cross-validation.
I think This is a good addition to the code to be submitted as it justifies the use of overfitting-resistant algorithms like Lasso or Ridge.